requirements.txt
requests==2.24.0
loguru==0.5.1
fire==0.3.1
zipstream-new==1.1.7